examples/Python3.11/torchvision-torchaudio-example/requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.5
filelock==3.18.0+ppc64le1
fsspec==2026.2.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3+ppc64le1
mpmath==1.3.0+ppc64le1
networkx==3.6.1
numpy==2.2.4+ppc64le1
Pillow==10.0.1+ppc64le1
requests==2.32.5
sympy==1.13.1
torch==2.6.0+ppc64le1
torchaudio==2.6.0+ppc64le1
torchvision==0.16.2+ppc64le1
typing_extensions==4.15.0
urllib3==2.6.3
